+++ DutchiePay/src/main/java/dutchiepay/backend/global/payment/dto/kakao/KakaoPayCancelRequest.java
@@ -1,4 +1,4 @@
-package dutchiepay.backend.global.payment.kakao.dto;
+package dutchiepay.backend.global.payment.dto.kakao;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
@@ -10,7 +10,9 @@ import lombok.Getter;
 @Builder
 @AllArgsConstructor
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public class KakaoPayCheckStatusRequest {
+public class KakaoPayCancelRequest {
     private String cid;
     private String tid;
+    private String cancelAmount;
+    private String cancelTaxFreeAmount;
 }
