+++ DutchiePay/src/main/java/dutchiepay/backend/global/payment/dto/kakao/KakaoPayReadyRequest.java
@@ -1,4 +1,4 @@
-package dutchiepay.backend.global.payment.kakao.dto;
+package dutchiepay.backend.global.payment.dto.kakao;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
@@ -10,10 +10,15 @@ import lombok.Getter;
 @Builder
 @AllArgsConstructor
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public class KakaoPayApproveRequest {
+public class KakaoPayReadyRequest {
     private String cid;
-    private String tid;
     private String partnerOrderId;
     private String partnerUserId;
-    private String pgToken;
+    private String itemName;
+    private Integer quantity;
+    private Integer totalAmount;
+    private Integer taxFreeAmount;
+    private String approvalUrl;
+    private String cancelUrl;
+    private String failUrl;
 }
